+++ src/ManagedModels/Events/ManagedModelArchived.php
@@ -5,7 +5,7 @@ namespace Thinktomorrow\Chief\ManagedModels\Events;
 
 use Thinktomorrow\Chief\Shared\ModelReferences\ModelReference;
 
-class ManagedModelUrlUpdated
+class ManagedModelArchived
 {
     public function __construct(public readonly ModelReference $modelReference)
     {