+++ src/Site/Menu/Events/MenuItemCreated.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Site\Menu\Events;
+
+class MenuItemCreated
+{
+    public function __construct(public readonly string $menuItemId)
+    {
+    }
+}